♻️ refactor : 초대된 대시보드 목록 컴포넌트화
