+++ src/@types/type.ts
@@ -63,3 +63,26 @@ export interface Dashboard {
   createdByMe: boolean;
   userId: number;
 }
+
+// 초대된 대시보드
+export interface Invitation {
+  id: number;
+  inviter: {
+    nickname: string;
+    email: string;
+    id: number;
+  };
+  teamId: string;
+  dashboard: {
+    title: string;
+    id: number;
+  };
+  invitee: {
+    nickname: string;
+    email: string;
+    id: number;
+  };
+  inviteAccepted: boolean;
+  createdAt: string;
+  updatedAt: string;
+}
