+++ src/components/atoms/sideBar/sideBar.module.scss
@@ -3,7 +3,6 @@
   align-items: center;
   min-height: 100vh;
   height: 100%;
-  background-color: #333;
   padding: 20px 24px;
   border-right: solid 1px var(--gray20, #eee);
   position: fixed;
